refactor(mod.ts): handling dataURL in mod.ts
