+++ src/renderer/main/package.js
@@ -7,14 +7,13 @@ import path from 'path';
 import { execSync } from 'child_process';
 import twemoji from 'twemoji';
 import * as matcher from 'matcher';
-import setting from './setting';
 import createList from '../../lib/updatableList';
 import replaceText from '../../lib/replaceText';
 import unzip from '../../lib/unzip';
 import buttonTransition from '../../lib/buttonTransition';
 import parseJson from '../../lib/parseJson';
 import apmJson from '../../lib/apmJson';
-import mod from './lib/mod';
+import mod from '../../lib/mod';
 import { getHash } from '../../lib/getHash';
 import packageUtil from './packageUtil';
 import integrity from '../../lib/integrity';
@@ -52,7 +51,7 @@ function getDate() {
  * @returns {Promise.<object[]>} An object of packages
  */
 async function getPackages(instPath) {
-  return await packageUtil.getPackages(setting.getPackagesDataUrl(instPath));
+  return await packageUtil.getPackages(mod.getPackagesDataUrl(instPath));
 }
 
 /**
@@ -341,7 +340,7 @@ async function checkPackagesList(instPath) {
   }
 
   try {
-    await packageUtil.downloadRepository(setting.getPackagesDataUrl(instPath));
+    await packageUtil.downloadRepository(mod.getPackagesDataUrl(instPath));
     await mod.downloadData();
     store.set('checkDate.packages', Date.now());
     const modInfo = await mod.getInfo();
@@ -377,7 +376,7 @@ async function checkPackagesList(instPath) {
  * @returns {Promise<object>} - An object parsed from scripts.json.
  */
 async function getScriptsList(update = false, modTime) {
-  const dictUrl = path.join(setting.getDataUrl(), 'scripts.json');
+  const dictUrl = path.join(mod.getDataUrl(), 'scripts.json');
   if (update) {
     store.set('modDate.scripts', modTime);
 
@@ -779,7 +778,7 @@ async function uninstallPackage(instPath) {
       await setPackagesList(instPath);
     } else {
       await parseJson.removePackage(
-        setting.getLocalPackagesDataUrl(instPath),
+        mod.getLocalPackagesDataUrl(instPath),
         uninstalledPackage
       );
       await checkPackagesList(instPath);
@@ -1068,7 +1067,7 @@ async function installScript(instPath) {
     };
 
     await parseJson.addPackage(
-      setting.getLocalPackagesDataUrl(instPath),
+      mod.getLocalPackagesDataUrl(instPath),
       packageItem
     );
     apmJson.addPackage(instPath, {
